address typings and refactoring feedback
